renamed lib to "cp-connector"
