+++ cp-connector/pkg/controlplane/controlplane.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 	"github.com/keptn/go-utils/pkg/api/models"
-	"github.com/keptn/keptn/lib-cp-connector/pkg/logger"
+	"github.com/keptn/keptn/cp-connector/pkg/logger"
 )
 
 var ErrEventHandleFatal = errors.New("fatal event handling error")
