+++ cp-connector/pkg/controlplane/controlplane_test.go
