+++ cp-connector/pkg/controlplane/eventsource.go
@@ -5,8 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/keptn/go-utils/pkg/api/models"
-	"github.com/keptn/keptn/lib-cp-connector/pkg/logger"
-	natseventsource "github.com/keptn/keptn/lib-cp-connector/pkg/nats"
+	"github.com/keptn/keptn/cp-connector/pkg/logger"
+	natseventsource "github.com/keptn/keptn/cp-connector/pkg/nats"
 	"github.com/nats-io/nats.go"
 	"reflect"
 	"sort"
