+++ statistics-service/main.go
@@ -77,7 +77,6 @@ func main() {
 	apiV1.GET("/statistics", api.GetStatistics)
 
 	apiV1.POST("/event", api.HandleEvent)
-
 	router.Static("/swagger-ui", "./swagger-ui")
 
 	apiHealth := router.Group("")
